requirements-venv.txt
streamlit>=1.44.0
streamlit-drawable-canvas>=0.9.0
Pillow>=9.0.0
pandas>=2.0.0
altair>=5.0.0
numpy>=1.24
# Pin 1.6.x to match vendored NOMA joblib (avoids InconsistentVersionWarning at load).
scikit-learn>=1.6.1,<1.7
joblib>=1.3
soundfile>=0.12
librosa>=0.10
matplotlib>=3.8.0
opencv-python-headless>=4.9.0
git+https://github.com/mozilla-ai/fake-audio-detection.git

# Research assistant (Serp / News / Fact Check + Gemini + Telegram)
httpx>=0.27.0
google-generativeai>=0.8.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
